docs: add more documentation
