+++ pyrut/types/__init__.py
@@ -4,6 +4,6 @@ from functools import partial
 
 from ..pyrut import validate_rut_string
 
-Rut = Annotated[str, BeforeValidator(validate_rut_string)]
+Rut = Annotated[str, BeforeValidator(partial(validate_rut_string, suspicious=False))]
 
 RutNotSuspicious = Annotated[str, BeforeValidator(partial(validate_rut_string, suspicious=True))]
